fix(term): eliminate 3 image protocol unwrap panics
- kitty.rs: replace RgbaImage::from_vec().unwrap() with proper error
  propagation (2 sites). A malicious Kitty image protocol payload with
  mismatched width*height*4 vs data length would panic. Now returns
  descriptive anyhow error.

- iterm.rs: replace candidates.pop().unwrap() with unwrap_or fallback.
  When both scaled dimensions exceed pixel bounds, candidates is empty
  and pop() would panic. Falls back to original dimensions.

Validated: 254 term tests pass via rch.

+++ frankenterm/term/src/terminalstate/iterm.rs
@@ -85,7 +85,7 @@ impl TerminalState {
 
                     candidates.sort_by_key(|a| a.0 * a.1);
 
-                    candidates.pop().unwrap_or((width, height))
+                    candidates.pop().unwrap_or((width as usize, height as usize))
                 } else {
                     (width, height)
                 }
